BPE26-ref: extract lifecycle function to dedicated file
