+++ src/main.ts
@@ -1,0 +1,9 @@
+import { logDetectedBrowser } from "./features/i18n";
+import { launchPlume } from "./features/lifecycle";
+
+(() => {
+  "use strict";
+
+  logDetectedBrowser();
+  launchPlume();
+})();
